Add: StudentPayment Flow

+++ src/main/java/org/example/labeebsystem/Service/CourseScheduleService.java
@@ -4,17 +4,21 @@ import lombok.RequiredArgsConstructor;
 import org.example.labeebsystem.API.ApiException;
 import org.example.labeebsystem.Model.Course;
 import org.example.labeebsystem.Model.CourseSchedule;
+import org.example.labeebsystem.Model.Teacher;
 import org.example.labeebsystem.Repository.CourseRepository;
 import org.example.labeebsystem.Repository.CourseScheduleRepository;
+import org.example.labeebsystem.Repository.TeacherRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
 public class CourseScheduleService {
 
     private final CourseRepository courseRepository;
+    private final TeacherRepository teacherRepository;
     private final CourseScheduleRepository courseScheduleRepository;
 
 
@@ -22,15 +26,24 @@ public class CourseScheduleService {
         return courseScheduleRepository.findAll();
     }
 
-    public void addSchedule(Integer courseId, CourseSchedule schedule) {
+    public void addSchedule(Integer teacherId, Integer courseId, CourseSchedule schedule) {
 
         Course course = courseRepository.findCourseById(courseId);
         if (course == null)
             throw new ApiException("Course not found");
 
-        if (course.getCourseSchedule() != null)
-            throw new ApiException("This course already has a schedule");
+        Teacher teacher = teacherRepository.findTeacherById(teacherId);
+        if (teacher == null)
+            throw new ApiException("Teacher not found");
 
+        if (!course.getTeacher().equals(teacher))
+            throw new ApiException("You are not the teacher for this course");
+
+        Set<CourseSchedule> courseSchedules = course.getCourseSchedules();
+        for(CourseSchedule cs : courseSchedules){
+            if (isOverlapping(cs, schedule))
+                throw new ApiException("Cannot Overlap Schedule");
+        }
 
         if (schedule.getStart_date().isAfter(schedule.getEnd_date()))
             throw new ApiException("Start date must be before end date");
@@ -38,10 +51,10 @@ public class CourseScheduleService {
         if (schedule.getStart_time().isAfter(schedule.getEnd_time()))
             throw new ApiException("Start time must be before end time");
 
+        schedule.setAvailability("Available");
         schedule.setCourse(course);
         courseScheduleRepository.save(schedule);
     }
-
 
 
     public CourseSchedule getScheduleById(Integer id) {
@@ -53,11 +66,24 @@ public class CourseScheduleService {
 
 
 
-    public void updateSchedule(Integer id, CourseSchedule updatedSchedule) {
+    public void updateSchedule(Integer id, Integer teacherId, CourseSchedule updatedSchedule) {
 
         CourseSchedule schedule = courseScheduleRepository.findCourseScheduleById(id);
         if (schedule == null)
             throw new ApiException("Schedule not found");
+
+        Teacher teacher = teacherRepository.findTeacherById(teacherId);
+        if (teacher == null)
+            throw new ApiException("Teacher not found");
+
+        Set<CourseSchedule> courseSchedules = schedule.getCourse().getCourseSchedules();
+        for(CourseSchedule cs : courseSchedules){
+            if (isOverlapping(cs, schedule) && !cs.getId().equals(schedule.getId()))
+                throw new ApiException("Cannot Overlap Schedule");
+        }
+
+        if (!schedule.getCourse().getTeacher().equals(teacher))
+            throw new ApiException("You are not the teacher for this course");
 
 
         if (updatedSchedule.getStart_date() != null && updatedSchedule.getEnd_date() != null && updatedSchedule.getStart_date().isAfter(updatedSchedule.getEnd_date()))
@@ -82,11 +108,34 @@ public class CourseScheduleService {
     }
 
 
-    public void deleteSchedule(Integer id) {
+    public void deleteSchedule(Integer id, Integer teacherId) {
         CourseSchedule schedule = courseScheduleRepository.findCourseScheduleById(id);
         if (schedule == null)
             throw new ApiException("Schedule not found");
 
+        Teacher teacher = teacherRepository.findTeacherById(teacherId);
+        if (teacher == null)
+            throw new ApiException("Teacher not found");
+
+        if (!schedule.getCourse().getTeacher().equals(teacher))
+            throw new ApiException("You are not the teacher for this course");
+
         courseScheduleRepository.delete(schedule);
     }
+
+    public boolean isOverlapping(CourseSchedule existing, CourseSchedule newSchedule) {
+        if (!existing.getDay().equalsIgnoreCase(newSchedule.getDay())) {
+            return false;
+        }
+        boolean dateOverlaps =
+                !(newSchedule.getEnd_date().isBefore(existing.getStart_date()) ||
+                        newSchedule.getStart_date().isAfter(existing.getEnd_date()));
+
+        if (!dateOverlaps) return false;
+        boolean timeOverlaps =
+                newSchedule.getStart_time().isBefore(existing.getEnd_time()) &&
+                        existing.getStart_time().isBefore(newSchedule.getEnd_time());
+        return timeOverlaps;
+    }
+
 }
